Reorganize Airspace types
